// src/core/constants/nature-name.enum.ts
export enum NatureName {
  ADAMANT = 'Adamant',
  BASHFUL = 'Bashful',
  BOLD = 'Bold',
  BRAVE = 'Brave',
  CALM = 'Calm',
  CAREFUL = 'Careful',
  DOCILE = 'Docile',
  GENTLE = 'Gentle',
  HARDY = 'Hardy',
  HASTY = 'Hasty',
  IMPISH = 'Impish',
  JOLLY = 'Jolly',
  LAX = 'Lax',
  LONELY = 'Lonely',
  MILD = 'Mild',
  MODEST = 'Modest',
  NAIVE = 'Naive',
  NAUGHTY = 'Naughty',
  QUIET = 'Quiet',
  QUIRKY = 'Quirky',
  RASH = 'Rash',
  RELAXED = 'Relaxed',
  SASSY = 'Sassy',
  SERIOUS = 'Serious',
  TIMID = 'Timid',
}